Changing the test project
Add DbDocumentReoisitoryStab

+++ TestProject.API/Controllers/DocumentsController.cs
@@ -25,6 +25,9 @@ public class DocumentsController : ControllerBase
     {
         try
         {
+            if (string.IsNullOrEmpty(userId))
+                return BadRequest("UserId is required.");
+
             var documents = _documentService.GetByUserId(userId);
 
             return Ok(documents);
@@ -41,11 +44,14 @@ public class DocumentsController : ControllerBase
     /// <param name="contractNumber">ContractNumber</param>
     /// <returns></returns>
     [HttpGet("contract-number")]
-    public IActionResult GetByContractNumber(string contractNumber)
+    public async Task<IActionResult> GetByContractNumber(string contractNumber)
     {
-        try
+        try //ToDo узнать, как лучше обрабатывать исключения.
         {
-            var documents = _documentService.GetByContractNumber(contractNumber);
+            if (string.IsNullOrEmpty(contractNumber))
+                return BadRequest("ContractNumber is required.");
+
+            var documents = await _documentService.GetByContractNumber(contractNumber);
 
             return Ok();
         }
@@ -62,10 +68,13 @@ public class DocumentsController : ControllerBase
     /// <param name="contractNumber">ContractNumber</param>
     /// <returns></returns>
     [HttpDelete("user-id-or-contract-number")]
-    public IActionResult DeleteByUserIdOrContractNumber([FromBody] DeleteModel model)
+    public IActionResult DeleteByUserIdOrContractNumber([FromBody] DeleteDocumentModel model)
     {
         try
         {
+            if (string.IsNullOrEmpty(model.UserId) && string.IsNullOrEmpty(model.ContractNumber))
+                return BadRequest("One of the parameters must not equal an empty string and null");
+
             _documentService.DeleteByUserIdOrContractNumber(model.UserId, model.ContractNumber);
 
             return Ok();
